+++ src/hoodie/store.js
@@ -5,17 +5,18 @@
 var hoodieStoreApi = require('../lib/store/api');
 var HoodieObjectTypeError = require('../lib/error/object_type');
 var HoodieObjectIdError = require('../lib/error/object_id');
-
 var generateId = require('../utils/generate_id');
-var config = require('../utils/config');
 
 var extend = require('extend');
 
-var utils = require('../utils/');
+var getDefer = require('../utils/promise/defer');
+var rejectWith = require('../utils/promise/reject_with');
+var resolveWith = require('../utils/promise/resolve_with');
 
 var localStorageWrapper = require('../utils/local_storage_wrapper');
 
-module.exports = function hoodieStore (hoodie) {
+//
+function hoodieStore (hoodie) {
 
   var localStore = {};
 
@@ -38,6 +39,8 @@ module.exports = function hoodieStore (hoodie) {
   var idleTimeout = 2000;
 
 
+
+
   // ------
   //
   // saves the passed object into the store and replaces
@@ -56,15 +59,15 @@ module.exports = function hoodieStore (hoodie) {
   //     store.save('car', undefined, {color: 'red'})
   //     store.save('car', 'abc4567', {color: 'red'})
   //
-  localStore.save = utils.toPromise(function (object, options, callback) {
-    var currentObject, event, isNew, key;
+  localStore.save = function save(object, options) {
+    var currentObject, defer, error, event, isNew, key;
 
     options = options || {};
 
     // if store is currently bootstrapping data from remote,
     // we're queueing local saves until it's finished.
     if (store.isBootstrapping() && !options.remote) {
-      return exports.enqueue('save', arguments);
+      return enqueue('save', arguments);
     }
 
     // generate an id if necessary
@@ -79,7 +82,8 @@ module.exports = function hoodieStore (hoodie) {
     if (isNew) {
       // add createdBy hash
       object.createdBy = object.createdBy || hoodie.id();
-    } else {
+    }
+    else {
       // leave createdBy hash
       if (currentObject.createdBy) {
         object.createdBy = currentObject.createdBy;
@@ -130,19 +134,22 @@ module.exports = function hoodieStore (hoodie) {
       delete object._$local;
     }
 
+    defer = getDefer();
+
     try {
       object = cache(object.type, object.id, object, options);
+      defer.resolve(object, isNew).promise();
       event = isNew ? 'add' : 'update';
-
-      callback(null, object, isNew);
-
       if (!options.silent) {
         triggerEvents(event, object, options);
       }
-    } catch (error) {
-      return callback(error, null);
+    } catch (_error) {
+      error = _error;
+      defer.reject(error.toString());
     }
-  });
+
+    return defer.promise();
+  };
 
 
   // find
@@ -153,28 +160,29 @@ module.exports = function hoodieStore (hoodie) {
   // example usage:
   //
   //     store.find('car', 'abc4567')
-  localStore.find = utils.toPromise(function (type, id, callback) {
-    var object;
+  localStore.find = function(type, id) {
+    var error, object;
 
     // if store is currently bootstrapping data from remote,
     // we're queueing until it's finished
     if (store.isBootstrapping()) {
-      return exports.enqueue('find', arguments);
+      return enqueue('find', arguments);
     }
 
     try {
       object = cache(type, id);
       if (!object) {
-        return callback({
+        return rejectWith({
           name: 'HoodieNotFoundError',
           message: '"{{type}}" with id "{{id}}" could not be found'
-        }, null);
+        });
       }
-      return callback(null, object);
-    } catch (error) {
-      return callback(error, null);
+      return resolveWith(object);
+    } catch (_error) {
+      error = _error;
+      return rejectWith(error);
     }
-  });
+  };
 
 
   // findAll
@@ -189,8 +197,10 @@ module.exports = function hoodieStore (hoodie) {
   //     store.findAll('car')
   //     store.findAll(function(obj) { return obj.brand == 'Tesla' })
   //
-  localStore.findAll = utils.toPromise(function (filter, callback) {
-    var currentType, id, key, keys, obj, results, type;
+  localStore.findAll = function findAll(filter) {
+    var currentType, defer, error, id, key, keys, obj, results, type;
+
+
 
     if (filter == null) {
       filter = function() {
@@ -201,7 +211,7 @@ module.exports = function hoodieStore (hoodie) {
     // if store is currently bootstrapping data from remote,
     // we're queueing until it's finished
     if (store.isBootstrapping()) {
-      return exports.enqueue('findAll', arguments);
+      return enqueue('findAll', arguments);
     }
 
     keys = store.index();
@@ -213,6 +223,8 @@ module.exports = function hoodieStore (hoodie) {
         return obj.type === type;
       };
     }
+
+    defer = getDefer();
 
     try {
 
@@ -249,11 +261,13 @@ module.exports = function hoodieStore (hoodie) {
           return 0;
         }
       });
-      return callback(null, results);
-    } catch (error) {
-      return callback(error, null);
+      defer.resolve(results).promise();
+    } catch (_error) {
+      error = _error;
+      defer.reject(error).promise();
     }
-  });
+    return defer.promise();
+  };
 
 
   // Remove
@@ -263,7 +277,7 @@ module.exports = function hoodieStore (hoodie) {
   //
   // when object has been synced before, mark it as deleted.
   // Otherwise remove it from Store.
-  localStore.remove = utils.toPromise(function (type, id, options, callback) {
+  localStore.remove = function remove(type, id, options) {
     var key, object, objectWasMarkedAsDeleted;
 
     options = options || {};
@@ -271,12 +285,17 @@ module.exports = function hoodieStore (hoodie) {
     // if store is currently bootstrapping data from remote,
     // we're queueing local removes until it's finished.
     if (store.isBootstrapping() && !options.remote) {
-      return exports.enqueue('remove', arguments);
+      return enqueue('remove', arguments);
     }
 
     key = type + '/' + id;
-
     object = cache(type, id);
+
+    // https://github.com/hoodiehq/hoodie.js/issues/147
+    if (options.update) {
+      object = options.update;
+      delete options.update;
+    }
 
     // if change comes from remote, just clean up locally
     if (options.remote) {
@@ -284,17 +303,24 @@ module.exports = function hoodieStore (hoodie) {
       objectWasMarkedAsDeleted = cachedObject[key] && isMarkedAsDeleted(cachedObject[key]);
       cachedObject[key] = false;
       clearChanged(type, id);
-      if (objectWasMarkedAsDeleted && object) {
-        return callback(null, object);
+      if (object) {
+        if (!objectWasMarkedAsDeleted) {
+          triggerEvents('remove', object, options);
+        }
+        return resolveWith(object);
       }
     }
 
+
+    //
     if (!object) {
-      return callback({
+      return rejectWith({
         name: 'HoodieNotFoundError',
         message: '"{{type}}" with id "{{id}}"" could not be found'
-      }, null);
+      });
     }
+
+
 
     if (object._syncedAt) {
       object._deleted = true;
@@ -306,14 +332,10 @@ module.exports = function hoodieStore (hoodie) {
       clearChanged(type, id);
     }
 
-    // https://github.com/hoodiehq/hoodie.js/issues/147
-    if (options.update) {
-      object = options.update;
-      delete options.update;
-    }
+
     triggerEvents('remove', object, options);
-    return callback(object);
-  });
+    return resolveWith(object);
+  };
 
 
   // Remove all
@@ -323,7 +345,7 @@ module.exports = function hoodieStore (hoodie) {
   //
   // when object has been synced before, mark it as deleted.
   // Otherwise remove it from Store.
-  localStore.removeAll = utils.toPromise(function(type, options, callback) {
+  localStore.removeAll = function removeAll(type, options) {
     return store.findAll(type).then(function(objects) {
       var object, _i, _len, results;
 
@@ -333,10 +355,9 @@ module.exports = function hoodieStore (hoodie) {
         object = objects[_i];
         results.push(store.remove(object.type, object.id, options));
       }
-
-      return callback(null, results);
+      return results;
     });
-  });
+  };
 
 
   // validate
@@ -451,8 +472,9 @@ module.exports = function hoodieStore (hoodie) {
   // clears localStorage and cache
   // TODO: do not clear entire localStorage, clear only the items that have been stored
   //       using `hoodie.store` before.
-  store.clear = utils.toPromise(function (callback) {
-    var key, keys, results;
+  store.clear = function clear() {
+    var defer, key, keys, results;
+    defer = getDefer();
     try {
       keys = store.index();
       results = (function() {
@@ -468,12 +490,13 @@ module.exports = function hoodieStore (hoodie) {
       }).call(this);
       cachedObject = {};
       clearChanged();
+      defer.resolve();
       store.trigger('clear');
-      return callback(null);
-    } catch (error) {
-      return callback(error, null);
+    } catch (_error) {
+      defer.reject(_error);
     }
-  });
+    return defer.promise();
+  };
 
 
   // isBootstrapping
@@ -874,10 +897,11 @@ module.exports = function hoodieStore (hoodie) {
   }
 
   //
-  exports.enqueue = utils.toPromise(function (method, args, callback) {
-    queue.push([method, args]);
-    return callback(null);
-  });
+  function enqueue(method, args) {
+    var defer = getDefer();
+    queue.push([method, args, defer]);
+    return defer.promise();
+  }
 
   //
   // 1. we store all existing data and config in memory
@@ -885,26 +909,26 @@ module.exports = function hoodieStore (hoodie) {
   //
   function moveData () {
     var oldObjects = [];
-    var oldConfig;
     var oldHoodieId;
 
     store.findAll().done( function(data) {
       oldObjects = data;
-      oldHoodieId = hoodie.id();
-      oldConfig = config.get();
 
-      hoodie.one('signin', function(newUsername, newHoodieId) {
-        for (var key in oldConfig) {
-          if (oldConfig.hasOwnProperty(key) && key !== '_account.username' && key !== '_hoodieId') {
-            config.set(key, oldConfig[key]);
-          }
-        }
+      if (! oldObjects.length) {
+        return;
+      }
+      oldHoodieId = hoodie.id();
+
+      hoodie.one('account:signin', function(newUsername, newHoodieId) {
         oldObjects.forEach(function(object) {
           if (object.createdBy === oldHoodieId) {
             object.createdBy = newHoodieId;
           }
-          store.add(object.type, object);
+          object = cache(object.type, object.id, object);
+          markAsChanged(object.type, object.id, object, {silent: true});
         });
+
+        triggerDirtyAndIdleEvents();
       });
     });
   }
@@ -930,5 +954,6 @@ module.exports = function hoodieStore (hoodie) {
     bootstrapDirtyObjects();
     delete store.bootstrapDirtyObjects;
   };
-};
+}
 
+module.exports = hoodieStore;