+++ src/utils/hoodiefy_request_error_name.js
@@ -1,10 +1,11 @@
 var findLettersToUpperCase = /(^\w|_\w)/g;
 
-exports.hoodiefyRequestErrorName = function (name) {
+function hoodiefyRequestErrorName (name) {
   name = name.replace(findLettersToUpperCase, function (match) {
     return (match[1] || match[0]).toUpperCase();
   });
 
   return 'Hoodie' + name + 'Error';
-};
+}
 
+module.exports = hoodiefyRequestErrorName;