+++ src/utils/local_storage_wrapper.js
@@ -1,36 +1,41 @@
 // public API
+var store = {};
 
-exports.setItem = function (name, item) {
+store.setItem = function (name, item) {
   global.localStorage.setItem(name, item);
 };
 
-exports.getItem = function (name) {
+store.getItem = function (name) {
   return global.localStorage.getItem(name);
 };
 
-exports.removeItem = function (name) {
+store.removeItem = function (name) {
   return global.localStorage.removeItem(name);
 };
 
-exports.clear = function () {
+store.clear = function () {
   return global.localStorage.clear();
 };
 
-exports.key = function (nr) {
+store.key = function (nr) {
   return global.localStorage.key(nr);
 };
 
-exports.length = function () {
+store.length = function () {
   return global.localStorage.length;
 };
 
 // more advanced localStorage wrappers to find/save objects
-exports.setObject = function (key, object) {
-  return exports.setItem(key, global.JSON.stringify(object));
+store.setObject = function (key, object) {
+  if (typeof object !== 'object') {
+    return store.setItem(key, object);
+  }
+
+  return store.setItem(key, global.JSON.stringify(object));
 };
 
-exports.getObject = function (key) {
-  var item = exports.getItem(key);
+store.getObject = function (key) {
+  var item = store.getItem(key);
 
   if (! item) {
     return null;
@@ -39,23 +44,23 @@ exports.getObject = function (key) {
   try {
     return global.JSON.parse(item);
   } catch (e) {
-    return null;
+    return item;
   }
 };
 
 function init() {
-
-  if (exports.isPersistent) {
+  store.isPersistent = isPersistent();
+  if (store.isPersistent) {
     return;
   }
 
   // if store is not persistent, patch all store methods
-  exports.getItem = function() { return null; };
-  exports.setItem = function() { return null; };
-  exports.removeItem = function() { return null; };
-  exports.key = function() { return null; };
-  exports.length = function() { return 0; };
-  exports.isPersistent = function() { return false; };
+  store.getItem = function() { return null; };
+  store.setItem = function() { return null; };
+  store.removeItem = function() { return null; };
+  store.key = function() { return null; };
+  store.length = function() { return 0; };
+  store.isPersistent = function() { return false; };
 }
 
 
@@ -70,8 +75,9 @@ function init() {
 // https://github.com/cappuccino/cappuccino/commit/063b05d9643c35b303568a28809e4eb3224f71ec
 //
 
-exports.isPersistent = function() {
+function isPersistent() {
   try {
+
     // we've to put this in here. I've seen Firefox throwing `Security error: 1000`
     // when cookies have been disabled
     if (!global.localStorage) {
@@ -97,8 +103,9 @@ exports.isPersistent = function() {
 
   // we're good.
   return true;
-};
+}
 
 // initialize
 init();
 
+module.exports = store;