+++ src/utils/promise/resolve_with.js
@@ -1,0 +1,9 @@
+var getDefer = require('./defer');
+
+//
+function resolveWith() {
+  var defer = getDefer();
+  return defer.resolve.apply(defer, arguments).promise();
+}
+
+module.exports = resolveWith;