+++ test/specs/hoodie/account.spec.js
@@ -2,7 +2,7 @@ require('../../lib/setup');
 
 var generateIdMock = require('../../mocks/utils/generate_id');
 var configMock = require('../../mocks/utils/config');
-var promise = require('../../../src/utils/promise/');
+var getDefer = require('../../../src/utils/promise/defer');
 
 global.stubRequire('src/utils/generate_id', generateIdMock);
 global.stubRequire('src/utils/config', configMock);
@@ -211,7 +211,7 @@ describe('hoodie.account', function() {
     }); // user is logged in as joe@example.com (hash: 'hash123')
     _when('user has an anonymous account', function() {
       beforeEach(function() {
-        this.signInDefer = promise.defer();
+        this.signInDefer = getDefer();
 
         // NOTE:
         // I do not understand, why we have to resetBehavior here.
@@ -511,7 +511,7 @@ describe('hoodie.account', function() {
             this.clock.tick(300); // do the delayed sign in
           });
 
-          it('should sign in', function() {
+          it('should send sign in request', function() {
             var args = this.account.request.args[1];
             var type = args[0];
             var path = args[1];
@@ -520,7 +520,7 @@ describe('hoodie.account', function() {
             expect(path).to.eql('/_session');
           });
 
-          _and('signIn successful', function() {
+          _and('sign in request successful', function() {
 
             beforeEach(function() {
               this.hoodie.store.findAll.defer.resolve([]);
@@ -530,12 +530,16 @@ describe('hoodie.account', function() {
               });
             });
 
+            it('should persist new username', function() {
+              expect(configMock.set).to.be.calledWith('_account.username', 'joe@example.com');
+              expect(this.account.username).to.be('joe@example.com');
+            });
             it('should trigger `signup` event', function() {
               expect(this.account.trigger).to.be.calledWith('signup', 'joe@example.com');
             });
 
             it('should resolve its promise', function() {
-              expect(this.promise).to.be.resolvedWith('joe@example.com', 'hash123', {moveData: true});
+              expect(this.promise).to.be.resolvedWith('joe@example.com', 'hash123', {});
             });
 
           }); // signIn successful
@@ -738,7 +742,7 @@ describe('hoodie.account', function() {
 
               _and('sign in fails with unauthorized error', function() {
                 beforeEach(function() {
-                  this.signOutDefer = promise.defer();
+                  this.signOutDefer = getDefer();
                   this.sandbox.stub(this.account, 'signOut').returns(this.signOutDefer.promise());
 
                   this.account.request.reset();
@@ -758,7 +762,7 @@ describe('hoodie.account', function() {
 
                 _and('sign out succeeds', function() {
                   beforeEach(function() {
-                    this.signInDefer = promise.defer();
+                    this.signInDefer = getDefer();
                     this.sandbox.stub(this.account, 'signIn').returns(this.signInDefer.promise());
 
                     this.signOutDefer.resolve();
@@ -1081,6 +1085,10 @@ describe('hoodie.account', function() {
 
       it('triggers movedata  event', function() {
         expect(this.account.trigger).to.be.calledWith('movedata');
+      });
+
+      it('does not trigger cleanup event', function() {
+        expect(this.account.trigger).to.not.be.calledWith('cleanup');
       });
     }); // signout succeeds
   }); // #signIn
@@ -1932,7 +1940,7 @@ describe('hoodie.account', function() {
 
             _and('sign in fails with unauthorized error', function() {
               beforeEach(function() {
-                this.signOutDefer = promise.defer();
+                this.signOutDefer = getDefer();
                 this.sandbox.stub(this.account, 'signOut').returns(this.signOutDefer.promise());
 
                 this.account.request.reset();
@@ -1953,7 +1961,7 @@ describe('hoodie.account', function() {
 
               _and('sign out succeeds', function() {
                 beforeEach(function() {
-                  this.signInDefer = promise.defer();
+                  this.signInDefer = getDefer();
                   this.sandbox.stub(this.account, 'signIn').returns(this.signInDefer.promise());
 
                   this.signOutDefer.resolve();
@@ -2094,7 +2102,7 @@ function presetUserDoc(context) {
   context.account.fetch();
   context.hoodie.request.defer.resolve(unconfirmedUserDoc(context.account.username));
   context.hoodie.request.reset();
-  var defer = promise.defer();
+  var defer = getDefer();
   context.hoodie.request.returns(defer.promise());
   context.hoodie.request.defer = defer;
 }
@@ -2121,7 +2129,7 @@ function with_session_validated_before(callback) {
         // now we have to reset the requestDefer
         this.account.request.reset();
 
-        defer = promise.defer();
+        defer = getDefer();
         this.hoodie.request.returns(defer.promise());
         this.hoodie.request.defer = defer;
       }
@@ -2151,7 +2159,7 @@ function with_session_invalidated_before(callback) {
         // now we have to reset the requestDefer
         this.account.request.reset();
 
-        defer = promise.defer();
+        defer = getDefer();
         this.hoodie.request.returns(defer.promise());
         this.hoodie.request.defer = defer;
       }