+++ test/specs/utils/promise.spec.js
@@ -1,6 +1,10 @@
 require('../../lib/setup');
 
-var promise = require('../../../src/utils/promise/');
+var isPromise = require('../../../src/utils/promise/is_promise');
+var resolve = require('../../../src/utils/promise/resolve');
+var reject = require('../../../src/utils/promise/reject');
+var resolveWith = require('../../../src/utils/promise/resolve_with');
+var rejectWith = require('../../../src/utils/promise/reject_with');
 
 describe('hoodie promises API', function() {
 
@@ -8,16 +12,16 @@ describe('hoodie promises API', function() {
 
     it('should return true if object is a promise', function() {
       var object = $.Deferred().promise();
-      expect(promise.isPromise(object)).to.be(true);
+      expect(isPromise(object)).to.be(true);
     });
 
     it('should return false for deferred objects', function() {
       var object = $.Deferred();
-      expect(promise.isPromise(object)).to.be(false);
+      expect(isPromise(object)).to.be(false);
     });
 
     it('should return false when object is undefined', function() {
-      expect(promise.isPromise(void 0)).to.be(false);
+      expect(isPromise(void 0)).to.be(false);
     });
 
   });
@@ -25,11 +29,11 @@ describe('hoodie promises API', function() {
   describe('#resolve()', function() {
 
     it('simply returns resolved promise', function() {
-      expect(promise.resolve().state()).to.be('resolved');
+      expect(resolve().state()).to.be('resolved');
     });
 
     it('should be applyable', function() {
-      var promise = promise.reject().then(null, promise.resolve);
+      var promise = reject().then(null, resolve);
       expect(promise).to.be.resolved();
     });
 
@@ -38,11 +42,11 @@ describe('hoodie promises API', function() {
   describe('#reject()', function() {
 
     it('simply returns rejected promise', function() {
-      expect(promise.reject()).to.be.rejected();
+      expect(reject()).to.be.rejected();
     });
 
     it('should be applyable', function() {
-      var promise = promise.resolve().then(promise.reject);
+      var promise = resolve().then(reject);
       expect(promise).to.be.rejected();
     });
 
@@ -51,7 +55,7 @@ describe('hoodie promises API', function() {
   describe('#resolveWith(something)', function() {
 
     it('wraps passad arguments into a promise and returns it', function() {
-      var promise = promise.resolveWith('funky', 'fresh');
+      var promise = resolveWith('funky', 'fresh');
 
       promise.then(function (a, b) {
         expect(a, b).to.eql('funky', 'fresh');
@@ -60,7 +64,7 @@ describe('hoodie promises API', function() {
     });
 
     it('should be applyable', function() {
-      var promise = promise.rejectWith('FUNKY!').then(null, promise.resolveWith);
+      var promise = rejectWith('FUNKY!').then(null, resolveWith);
       promise.then(function (error) {
         expect(error.message).to.eql('FUNKY!');
       });
@@ -71,7 +75,7 @@ describe('hoodie promises API', function() {
   describe('#rejectWith(something)', function() {
 
     it('wraps passed arguments into a promise and returns it as Error', function() {
-      var promise = promise.rejectWith('funk overflow!');
+      var promise = rejectWith('funk overflow!');
 
       promise.then(this.noop, function (error) {
         expect(error).to.be.an(Error);
@@ -84,7 +88,7 @@ describe('hoodie promises API', function() {
     });
 
     it('should be applyable', function() {
-      var promise = promise.resolveWith('wicked!').then(promise.rejectWith);
+      var promise = resolveWith('wicked!').then(rejectWith);
       promise.then(this.noop, function (error) {
         expect(error).to.be.an(Error);
         expect(error.name).to.eql('HoodieError');