+++ CHANGELOG.md
@@ -1,3 +1,55 @@
+<a name="0.9.6"></a>
+### 0.9.6 (2014-06-10)
+
+
+#### Bug Fixes
+
+* **remote:** trigger doc:change, not change:doc ([355bd064](https://github.com/hoodiehq/hoodie.js/commit/355bd0642185c517647fff3564cc3723a8baa223))
+
+
+<a name="0.9.5"></a>
+### 0.9.5 (2014-06-10)
+
+
+#### Bug Fixes
+
+* **store:** when object removed from remote, remove it locally instead of marking as _delete ([bcd656fe](https://github.com/hoodiehq/hoodie.js/commit/bcd656fe959e6e8c842c7e13d81ba864cc570170))
+* **utils:** fixed store.getObject / store.setObject behaviour ([a4102a69](https://github.com/hoodiehq/hoodie.js/commit/a4102a69db2f917b7db282d34c7b45a87888acec))
+
+
+<a name="0.9.4"></a>
+### 0.9.4 (2014-06-06)
+
+
+#### Bug Fixes
+
+* **account:** signin in from anonymous account now moves data correctly ([914dd515](https://github.com/hoodiehq/hoodie.js/commit/914dd515ee80fedda8afe92bffa1ff0b9437f3e5))
+
+
+<a name="0.9.3"></a>
+### 0.9.3 (2014-06-06)
+
+
+#### Bug Fixes
+
+* **sync:** trigger remove events for completed tasks ([a2d94fa8](https://github.com/hoodiehq/hoodie.js/commit/a2d94fa88c90669d2aee8998d8840a710d919781))
+
+
+<a name="0.9.2"></a>
+### 0.9.2 (2014-06-05)
+
+
+#### Bug Fixes
+
+* **account:** hoodie.account.signUp() to persist username ([5d58bf08](https://github.com/hoodiehq/hoodie.js/commit/5d58bf088159d907473151073580c5ec7cc2845d))
+* **hoodie:** pass lib & utils to hoodie.extend callback ([92cc90e8](https://github.com/hoodiehq/hoodie.js/commit/92cc90e8c1be7cca3cd882ab201e6cd5babaabfe))
+
+
+#### Features
+
+* **account:** 'passwordreset' events now pass username ([3048451c](https://github.com/hoodiehq/hoodie.js/commit/3048451cf4517fcf0f193d31e3af72cd09af15c6))
+
+
 <a name="0.9.1"></a>
 ### 0.9.1 (2014-05-28)